Investigation on better console handling
I'm tired because we are thursday, but I wanted to try something. termios doesn't exist on Windows, but I found something pretty similar that will allow me to better handle the user input (your days are counted scanf).
The code is ugly for the moment, but now I have the possibility to move the cursor by one with the Right arrow, and back with the left arrow. Which is awesome because I will be able to develop something more user friendly for their inputs (game settings, and pegs values).