Terminal control advancement
I have looked the Windows documentation to create a menu for our game.
Currently, the game name is displayed on boot, and then the user can choose between 3 different lines (with arrows up/down ) what he wants. On enter, it will be executed depending of what the user chose.
The code is still ugly, modifying the terminal settings exploded scanf afterwards when starting the game, but who cares ? We have a nice menu !
Next steps would be to chain different menus ( main -> settings -> start game as an example ) and then redesign the game display.
Hyped !