Game menus addition
In order to create a proper flow for the game menus, I've created some structures and the associated file to support a logic with multiple linked menus.
For the moment, I've replaced the old code by the new one, I need to improve some elements (add the title), and others things, but the main menu begin to look good.
Next step, would be great to have a nice settings menu for the game. ( shared between singleplayer and multiplayer )