Key inputs rework
- The previous Enum definition was quite incorrect with the enum count because the values weren't linear.
Now, the KeyInput enum is linear from 0 to 70 with some helpers inside to delimite categories.
The association between the enum and the key values has been made internally, alongside the key name (mainly to help futur debugging).
- Also added a bunch of useful functions to convert the raw input to an enum, and identify easily which type of key has been pressed (letter, number, ...)
- Modified game_menus.c to use the ney KeyInput API.

The code is cleaner on this side, happy with the result ( could have used a define in the implementation instead of copypast the same struct definition everywhere though. If we want to change a field, it will require more updates... ). Later perhaps, not urgent.