External terminal and resize handling
This experimentation was so painful to be honest.
Having a dedicated terminal wasn't the hardest part ( We now have a console for our game which is nice instead of using the one where the exe has been started ), but handling the resizing of the window was... ugh.
We can't have a fixed width, so we need to pay attention to the current window size. If too small, display an appropriate message (aka need bigger window ), and otherwise display our game. I have now a combination of escape sequences thata llow us to :
- Avoid having a scroll bar / display history
- Glitches if we are resizing rapidly the window
- Empty lines depending of the cursor position.
Now that we have a correct handling of the console for the title display, we need to hook the rest of the game in the same pattern.