Hardcoded menu from new display mecanism
New that we have our own terminal, I have made some hardcoded code ( please don't judge it, I hate it myself but I want to test some things. Consider it as a POC that we will be rewritten if it works as expected ).
We have now the title, and some menu tile that are displayed, with the screen rewritten either when a valid input is received, or when we need to rewrite the screen (aka screen has been resized).
With the new sequence that we use to clean the screen, the flickering has re-appeared. I would need to find a solution to combine both needs : avoid flickering but also handle the resize accordingly.