Multiples attempts and testing
Yeah that commit is quite messy, I've tried several approaches to either create the menu of the game, or handle the console. The double buffering solution was interesting to use, but I read that we should use the alternate screen buffer of stdout instead of creating a second screen buffer. The code has moved up a bit as well, I'm not gonna describe it all.
the console.h/.c file is the file that will handle all the console stuff, so this is the important file of the day.
This commit is just here to keep a trace of all my attempts before cleaning up all the ones that I don't like.