Console printing with virtual terminal sequences
INstead of using the Windows API, I've added more utilities by using the virtual terminal sequences ( color, cursor position/display, ... ).
The code has been cleanup a little bit, and now the main function is displaying the boundaries of the console with the draw lines characters set.
We could try to either have the MASTERMIND title centered in the console now, and handle the resizing to adapt the display accordingly.
Once done, we will be able to setup the menus and game layout.