Hardcoded board (Testing design)
I've hardcoded the display of the potential game board of our Masterming games. I like it so far, I'll sleep and see if I like it as much tomorrow.
If validated, we will need to de-hardcode the function in order to be relative from a position / using actual data ( pegs colors, ... ) as well.
That's some nice progress here !