Non-hardcoded Boardgame generation
The boardgame generation is now made depending on the upper left coords given, and also adapt depending of the number of pegs per turn, and the number of turns.
The code has been cleaned up a little bit, but the generated board is just an empty version.
We don't have any data (yet) from an actual game. Next step, we should setup the Pegs/FeedbackPegs structures in order to display correctly

I've also updated the resize mechanism. Instead of resizing at each event received, we will now first check all events received during that frame, and then resize at the end if necessary.
It prevent having a "queue" of resize events as the redraw takes some time.
We could also think about a solution to avoid clearing the entire screen, but only the important part ? It would be smoother for the user.