Stable 30 fps !
Hurray !! After having retried and retried with Sleep, nanosleep, SleepEx, QueryPerformanceCounter, QueryInterruptTimePrecise and other functions, I've finally succeeded to keep a steady 30 fps ( and not ~26fps because sleep functions have a precisions of ~15ms ) by using CreateWaitableTimerExW !
While using it at the end of the loop like the other waiting function doesn't really work, setting the timer at the beginning of the loop with the wanted capped fps allow us to just wait INFINITE at the end of loop on it. If we were too fast, it will wait on the object. If we missed some frames, it will skip the Waiting.
The code is a freaking monster as I didn't cleanup my attempts, and my broswer has 239 windows. will be a pleasure to close all of them.
Now that's issue is fixed, I'll clean up the code, and perhaps create a little wrapper to better handle these nanoseconds shenanigans.
Happy sleep to me !