Consume inputs into mastermind implementation
- This reduce the code in the main function, and also allowed us to move the draw code into the mastermind file.
- Reduced the FPS impact by removing the entire redraw, to only redraw the needed things.
The only thing that is not 60FPS constant now is the restart as there is still a lot of IO operations ( char by char ). Filling a buffer, and then refresh at the end of the frame could be better to mitigate this issue.
And if the buffer doesn't resolve the FPS and flickering issue, we will return to our double-buffering solution (it's Windows specific though :( )