Add a refresh function to display the update
Instead of doing X IO calls with wprintf, we will now accumulate the updates into a buffer by calling console_draw(), and then only doing one IO operation with console_refresh().
It works great, and now a resize is ~45fps instead of the previous ~9fps. The other updates doesn't fluctuate the fps. The flickering is still present, we might want to still do the double-buffering (but only on resize to not change the screen buffer every frame).