Pause and design POC
I've drank my cup of tea outside and thought more about the overall design of the game. The board being Horizontal is not very intuitive... So we could have a big vertical board and a summary top right to keep everything in the same window if the board is bigger than the screen. I've put an hardcoded version that I like, so I'll make it work with the data that we have, and attach to the right, so we move on resize.