Widgets idea !
In order to handle the different elements of the screen, it could be nice to separate them by widget.
The summary part could be a Summary widget where a special function will draw the borders
I've implemented a Countdown (which is nice by the way), where the menu is generated by these functions, and afterwards the specific Countdown function display the necessary things.
There are a lot of things to improve in the design, but I think it would be good, a menu could be a widget ( like the pause menu, with custom border color, ... )