Widget FPSBar and delay on resize
- Moved the FPS / ms draw calls that where on the main loop in a dedicated widget, FPSBar.
There is a init/uninit function that hook it up in the widgets manager thing.
There is a dependency with the FPSCounter system, we could perhaps move the widget init inside to hide that part of the implementation.
If we want to hide the widget, perhaps we could at least make the WidgetId public for everyone. This way, we could say "hide the widget X" to the manager, and he will look on its hooked widgets.
- I've also added a little delay before the redraw on resize (200ms). This way, we are preventing the game to redraw every frame and dropping the framerate. Here we have only one spike, when the entire thing is redrawn -> double buffering could be the solution here