Global widget draw borders function (WIP)
- Working on a global version of drawing borders for our widgets, based on the implementation of the one made for the summary board.
Still in progress, but the borders are dashed and yellow if the required size is not met. It's working nicely, but I fear that it won't be clear for the user that it needs to expand the console.
I will try a version with a dashed arrow at the end (just before the limit)