Main board WIP design and small code cleanup
- Created a bunch of files :
mouse.c/.h to handle the mouse position, console_screen.c/.h for the screen resizing, and the future widget for the main board.
Now we would need to hook some functions into the console_screen to be called on resize. This way, we will avoid checking each frame for an update.
- The main board has a final design, I will now be able to begin working on the Board display in its own widget.