A rework in progress, but for the best !
After a good night of sleep, I have decided that the current design for the game (widgets and borders) was garbage.
Displaying the - - - and empty the widget on truncation was a good experimentation, but it turns out that it's not really a nice user experience.
So I'll rework that little bit by little bit. The screen size part has been updated to be responsive to the size update, and is more explicit than before to tell the user to increase the screen size.
- The framerate widget has been cleanup
- The Screen Size widget has been added, and clean up (was the old code in the main.cpp )
- The mouse position on top has been removed, will see if it's still needed in the future
- Added a new struct Rect in order to handle rectangles with collisions. With that method, it's easier to get a corner of the rectangle, to know if a point is in the rectangle (which works great with the mouse hovering) .
- A rework on the way will also be to separate the incoming buffer from the buffer that we will write in the end in the console. We need to store the currently displayed buffer and the incoming one, compare the two and generate the difference, and update the screen with that difference.
We will need to think on the resize part, where we need to rewrite the part that have been erased by being shrunk.
I think it will be the last modification of 2023, so Happy new Year everyone :)