MOOORE BUTTONS !
Added more buttons in the top ( abandon game ), and also added a lot of buttons below the game board, in order to have all the possible actions that the user can do to be handled with the mouse or keyboard. The part with the mouse isn't documented in it, but good enough.
- Each button can choose its own color, which means that the reset is in red, and the confirm in green. Perhaps we want the abandon in red as well ?
- We will need to disable/enable buttons depending of the state of the game. You can't confirm the turn on a turn that isn't valid for example.
- wprintf explodes on the first draw ( so I've disabled the board summary for now ). We are writing too much chars ( >4096) and printf functions doesn't suppor tthat well. We will need to print block of 4096 chars to avoid this behaviour.