Mastermind game reworked implementation (WIP)
- Now that we have a system of callbacks to bind the UI to the game, I've added a similar system to bounce back a game update into the UI.
For example, if the UI send 'A' on a click for Abandon Game, the game will consume that input, update the game accordingly, and then send a "StatusUpdate" update that widgets will be able to take in account to adjust the UI, and for our edge case, make the game buttons disappear and disable the abandon button (we can't abandon a game that is not in progress ).
This system works well with the buttons, and now that we have a simpler and much efficient base for the mastermind game, our next step will be to sync the game with the main boatd UI. ( and the next step after that would be on the board summary ).
- Cleanup has been made and files that weren't used anymore have been deleted from the project in order to avoid keeping too much dead code.
- The Settings file has been created, in order to store the values that we will display to the user in a future settings widget. The user will be able to choose its fps cap, and game settings. The game settings update won't be effective on the current game, the user will need to create new game (and this should be marked in the menu ).