Main board development
- The main board needed some love since a long time, so I've began to really work on it. The pegs and pins display is now based on the current state of the mastermind game.
The main feature (which works correctly for now) is the history. With U/D keys or the scroll in the mouse, it's possible to navigate the board (we are only displaying 4 rows max on a higher numbers of possible turns ).
When the history is on top, the title is displayed, and when the history is down, we have the solution ( a title for the moment, easier to test the conditions ).
- Added a new callback to give the possibility for a widget to consume an input. The input will be consumed before the mastermind game. It's good for now, but in the future we should have a better way to handle the key bindings. ( low prio ).
- Next step is to finish the title display ( hardcoded size at the moment ), and a solution as well at the end. In order to test the peg display, abandon game should :
- reveal the solution
- display the solution to the user by auto-scrolling to the end of the board.