Color Selection Widget (WIP)
- Created a first attempt of having the list of colored pegs in order to choose the peg from.
That first attempt draws the pegs in the bottom right, while displaying with a bigger circle the current selected one. While it seemed good on paper, it doesn't seem really instinctive with a keyboard. Perhaps we could attach the different pegs into the board directly ( like on the right ), in order to centralize that in only one widget.
- Jumping from the board to the pegs selection (and vice-versa) and cumbersome. We need to find an elegant way to handle that for the keyboard.
On the mouse, it could be fun to implement a drag-and-drop mechanism (where the first click attaches the peg, and the second drop it). From a mouse user perspective, it would be great to use.
However, we need to work the keyboard feelings here.
- On a side note, I've removed the pins from the board if the feedback hasn't been generated yet. The current turn has also been changed into yellow ( the past one in white and the future in grey ) in order to find your way on your current turn more easily.