Board Summary Improvement
The board summary ( renamed simply Summary ) was only hardcoded. Now its size and position is linked to the number of turns and pegs of the mastermind game config.
- While developing this "widget", I was thinking about transitioning from widget to components. It seems more relevant than widgets with how they have evolved.
- Added a new rect_ function to draw the border of the given rect. It's based on the code in widget_utils, and reused for the summary component. SO we have the name + borders displayed.
- We are in the same state than the board here, the next step would be to be able to add a peg to the board, and confirm the turn.
Otherwise, we could begin to rethink how we are displaying things in the console (check the previous commits ). We also need a "DARK" version ( like "BRIGHT" ) except is not handled the same ( \x1b[2m ), it's not a color.