Console Screen handling rework (WIP)
So this is the beginning of a rework that I already brought on my previous commits, where the current console implementation is a little cumbersome, and not very easy to use. There is also the bright/dark version of the color that isn't handled at all and I needed to directly add the virtual sequence keys in some widget, which is not correct.
- Firstly, I've moved the console_screen file into a new console folder.
- The colors will be stored in a bitmask, to change it more efficiently ( there is also a default mask for the black background + white foreground ). A similar logic has been made for the style : bold, underline, ...
The shade of the color (bright/dark/default) will also be easier to manipulate.
This combinaison of elements is called attributes, and stored for each cell that composes the console screen. So in the future, we will have in memory the whole screen with the current character + its attributes.
This console screen will be doubled for the current frame, and the incoming one. This way, all modifications asked by the game will be only reflected in the incoming buffer (no virtual sequence yet). This is only after that we will generate the buffer that allows us to transform the current frame to the incoming frame, and print it to the console.
This way, all characters that haven't changed between the 2 frames won't be redrawn, avoiding a flickering effect.
This commit is only the begin of the implementation, there is no usage yet.