Console Screen handling rework (Part 2)
Still working on the console screen rework in this commit.
Here the implementation is enough developed to have a MVP : the framerate and screen size components are displayed on screen using the new implementation.
The code is quite a mess, but it's promising on how it will render on screen. We are indeed avoiding rewriting characters that haven't changed, so the flickering is less present.
However, the resize is not well handled for the moment. When the screen is reduced, the characters are erased. if the screen is extended again, we should rewrite these affected columns/lines.
It's far from being finished, but the new way of handling the style/color is cleaner, which is one of the reasons of this rework. Very happy about that !