Console Screen handling rework (Part 3)
After some thoughts on what I'd done on the new implementation, I was not satisfied on how the colors were handled. Bitmasks are useful, be here he would allow a color to have multiple colors enabled at the same time, which shouldn't be possible.
So instead, I've kept the binary values of the colors ( from 0b000 to 0b111 ) in order to have the background and the foreground colors in the same byte. With 2 unused bits, I've added the brightness option for both colors. This resulting to a nice byte entirely filled with useful information. Coupled to the byte of attributes, our Style has a size of 2 bytes, which is relatively small.
- I've also corrected a major (and dumb) error in my previous values declaration, I was using hexadecimals ( 0x ) instead of binary ( 0b ).. So the old version was much bigger than this improved version.
It's also simplify a lot of logic as we can directly used the color value by shifting the bits when necessary.
- The way of writing the terminal sequences in the buffer has also been reworked to simplify the usage. Now the code is much smaller, and also easier to read.
It's still work as expected on the console, so it's a good time to submit that part of the changes before diving deeper.