Console Screen handling rework (Again)
Okay, I know that it begin to make a lot of commits for that rework alone. But I haven't got much time at the end of the day now that I'm working again.
So today, I decided to make a big move and name the screenpos screenpos_deprecated in order to improve the usage of that element.
Furthermore, I've considerably reduced the size of the screen data, which has now the size of ~14Kb, instead of more than ~40Kb before.
The double screen frame has been removed to only keep one instance of the screen buffer, and use one of the unused bit of the Flags byte to know if a position of the screen needs to be refreshed.
This way, we can effciently rewrite the information at a position, and then know which positions needs to be rewritten on refresh.
In other words, the overhead of rewriting multiple times the same position doesn't exist anymore.
- A separation has been made for the real screen size, and the actual game area, which can't be bigger than 120x30.
- The struct Style has been renamed Properties because the flag "needs refresh" isn't part of the style.
- The time typedef have been renamed to be more explicit (despite being longer now ). e.g. nsec is in nsecond, and min is minute.
The result is still correctly displayed, so next time it will be time to acknowledge if this is the path we want to take, and cleanup the implementation in consequences.