Terminal & Components refactoring advancement
Today is a great day.
This is far from finished, but the result that I hoped getting from this refactoring begin to show.
- I've began to rename all console_* things into terminal_*. Perhaps it's because of my job, but console always make me think about XBOX & SONY consoles, so terminal it is.
- The code under the folder terminal/ is been divided into several files instead of one :
character / screen / sequence and style. Each have a clear scope, and simplify the implementation code behind because it makes things more explicit. The only thing that I'm still unsure is if I keep the term_screen instead of term_. It's just annoying for the user to search if it needs to add screen or not.
- Widgets continue their transformation into components, and I'm becoming satisfied with this new name.
- The temporary code for the mouse position has disappeared in main and moved into its own component.
Let's talk about the main feature :
With the rework of term_write and term_refresh, we have reduced the size of the buffer that we are sending to IO by optimizing the term sequences changes. By playing with our "needs refresh" bitflag, we are now flagging all the characters that have been erased because of the resize, in order to rewrite them once the terminal size increased again. The feeling is pretty smooth on the user size, but we need to experiment more once the screen more filled. Perhaps we would need to add a little delay before writing them again after a resize (and avoid for example writing a character from the line 1 into the line 2 if the user moves the size non-stop ).

While the rework isn't yet finished, the current result is already promising on our quest to keep a steady 60fps.