Board game improvement
I've changed the design of the main board. The blue board has been deleted for something cleaner (and less bloated). However, the display isn't final yet, there are some elements that are not correctly displayed anymore (solution and pins). It's the same style as the rest of the components so there is more harmony for the eye.
I've also experimented some buttons inorder to have the name + the associated keybind. I have 2 choices, I think that I will use them both depending of the situation instead of making a choice.