Button style update
After discussion with a friend ( the first feedback received on the game ), I agree that the buttons aren't the most explicit on the associate keybind, and quite heavy on top of the screen.
So I've switched the "Development build" mention to the top and brought the buttons at the bottom of the screen.
On top of that, the buttons are now displayed on 2 lines, with the first being the button name, and the second the associated keybind.
I was not very consistent before as some keybinds were between [ KEYBIND ], or some other had the button name with it.
Now we can understand that having [ ] is related to keybindings.
I've also change a little bit the style colours.

I'm overall happy with these changes, now we need to focus more our attention on the board and its mouse handling.