Reduce refreshes that redraw everything (small components)
So I've raised to 120 the capped framerate of the game to better see the fps variation depending of what I'm doing with the game.
Calling a refresh that redraw everything each time is consuming, and reduce the fps by a lot.
The improvement we could do is having 2 required callback for each component : enable and disable.
This way, a component will be able to draw itself completely on the enable ( and clear on the disable), but the other refresh updates will be smaller : A button will update itself on hovering for example, without calling a refresh on the component is in.
I know that's it's not the priority, but I didn't have much time this evening to do something that would have required more brain to work.