Piece Peg/Pin rework (WIP)
The logic for the different pieces of the game ( pegs and pins ) were spread in multiple files, and drawn in different files as well.
Here is a first attempt to have a unique byte for the pegs/pins, with flags and all to characterize the different pieces of the game. It also makes all the drawing at the same place, which is cleaner. However, there are a lot of bitwise operations, and not a lot of helpers yet.
The game compiles but there are a few changes to make to repair all the feature.