Enable/Disable callbacks
In this commit, I'm doing even more cleanup in our small components (framerate/mouse/screen size/game buttons) and summary in order to reduce the refresh size, and implementing the enable/disable callbacks while removing the refresh callback.
The summary has also been adapted to be played with 4, 5 or 6 pegs. The size of the component doesn't change, but the display inside is centered depending of the number of pegs played.