Next week, next idea (GameEvents / UIButton)
Okay so here we go with some improvements that I thought about ( I'm still not sure how to compose all the bricks together though ), but the goal would be to transform the current MastermindEvent to a wider GameEvent ( which could contain QUIT request, ... ). The goal here is to drive the game modifications by events, and reducing the dependences between components.
For example, the button QUIT will emit the key event associated to the keybind quit ( yes I'm begin to add a keybinding layer, perhaps to be able to customize it in the future ). If this keybind is catch by the gameloop, it will trigger a quit event request, that will set the main loop to false.
This layer of key -> events seems to be a proper way to handle communications and updates between all the components and the game itself.
We will see how it goes once I'll apply this logic on more components and sub-systems.
- Another change is the ui.h header (only used for buttons for now ), where the buttons logic has been given to a common file instead of being duplicated in game_buttons, pegs_selection and board. The last 2 components haven't been updated yet, but it's on the agenda.