Replacement of MastermindEvent to Events - WIP
In this commit, I'm pursuing the transition from the MastermindEvents and the call to fucntions to being exclusively Events based.
The goal would be to create a more centralized way to have the keybinding -> event -> updating game / create others events maybe.
The transition is still a mess in the code, but it's in WIP.
On the question "But why these changes ?", it's because the interaction between the PegSelection component and the board when you want to add a peg to the board with the mouse was complex / impossible without creating hard dependencies between components, and exposing everything. If we are sending events for each modification instead, we could just sync the UI with that.