Generalize input subscriber for event handlers
The event handler was taking a Widget * as first parameter for our subscriber. However, there is not only the widgets that would you like subscribe on the events : we may want to subscribe on the INPUT event the check for quitting the game.
I've also added a return value for the event callbacks now, in order to specify if the event can be propagated or stopped. It'll useful for the input event for example, if one specific widget/UI widget has consumed the input.