Widget bottom navigation and inputs treatment changes
The inputs received by the player (either by the keyboard or the mouse click) will be now wrapped under the event "USER_INPUT". This allow our bottom navigation bar to subscribe to this event, and send it to its buttons to see if there is a match.
These buttons have now a callback "OnPressed" that will  (soontm) generate a Request, not an event.
These Requests is the next step, in order for the game to update itself. Then on update events could be triggered and we will have our complete loop :
UserInput -> Request -> Event -> UI update