Requests implementation
The requests are now available to use !
Our first usage is the Quit button, that will send an EXIT_APP request. The gameloop will then process that request, and stop the main loop. The request for the NEW_GAME has also been developed, so the mastermind logic is restarting the game. However, we don't have any visual yet to confirm that in our new widgets, so next step : Move our summary widget into the new world. We will be able to experiment with our Request->Event->UI update loop a bit more.