Improved the peg update events
Instead of having general events like "hide/show solution" where the widgets needs to understand that we need to hide/show all pegs of the solution, now the mastermind will send an event PEG_HIDDEN/REVEALED, with the turn/index and gamepiece. This way, it will be easy to adapt the code if we need to hide some pegs of the board ( not just the solution ).
- The arrows and backspaces are also taken into account now, thanks to the board widget. This board doesn't display anything yet, as we are first working with the summary to get things right.
The next thing to handle would be the turn validation and the feedback display.