Feedback pins and game completion
- The feedback pins has been re-added when the turn has been confirmed.
Perhaps we should send an event "Turn Valid / Turn Invalid" depending of the state of the turn. This way, we could enable/disable the "Confirm Turn" button. We should also have some tracking in board to enable/disable the reset button.
- The Mastermind can be completed now with the Summary widget. The only thing is that we can't see the current position in the row. I thought of adding it, but this isn't the goal of this widget.
Now the focus will be on the board widget.