Peg drag readded, and fixed
- The dragging mechanism has been readded, and developed in a separate widget ! Which is cleaner than having it in the PegSelector widget.
I've also fixed a bug that we had with that feature before when we are hovering a button, we are overriding the button color with our save characters. So we could end up with a little touch of yellow and the rest white.
I believe that the next would be the board interaction  with our mouse and the selected peg.