game piece (pegs/pins) rework
The current gamepiece is a byte composed of some flags (hidden, empty ) and the color of the pegs. By being a single byte, it was adding complexity for nothing as we would always check for the flag empty, but the color was kept, and some visual bugs could be encountered.
Now, EMPTY is one of the possible Id ( Alongside the differents colors ), and we have a boolean for being hidden or not.
I have changed the gamepiece into a struct Piece / struct Peg / struct Pin everywhere, and the result is cleaner.